requirements.txt
fastapi
uvicorn
openai
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1